Improve event subscription (#423)
This should improve events reliability when subscribing to events with FQDN containing localhost or network aliases or when there are multiple network interfaces (#315). The situation is especially improved when using an AttributeProxy for the event subscription.
Limitation: in case DeviceProxy is used for event subscription, client might get an incorrect event name value in the event data (containing a TANGO_HOST  hostname slightly different than the one used during the FQDN subscription or in the env. variable) in the event callback if short domain name is used in TANGO_HOST environment variable e.g. hzgc103k -> hzgc103.desy.de