zmqeventconsumer/zmqeventsupplier: Make it compatible with all stock cppzmq versions
Depending on the cppzmq version from https://github.com/zeromq/cppzmq the
operator void* is either implicit or explicit.

We could now add a feature check in cmake to check for that and use code
paths for each variant.

But we could also be really pragmatic and just use a unconditional
static_cast<void*>.

Idea and initial patch by Michal Liszcz.