Travis Tests: Avoid code duplication
We currently have run scripts for each supported debian version. These
run scripts are mostly very similiar.

So it makes sense to use a common script.

The differences are if we use COVERALLS support which we can
support via environment variables.

We also reduce the number of jobs in preparation for adding debian 10
(buster) support. We only keep the debug jobs of debian 9 and lower.