Add data_type_to_string for outputting data types
We used to suggest to users to use

CmdArgType[type]

for getting a string representation of the type.

But that fails horribly, i.e. with a segfault, if you pass DATA_TYPE_UNKNOWN
into it.

DATA_TYPE_UNKNOWN has a value which is way above the last used entry in
order to be future proof.

The newly added function data_type_to_string gets the string
representation and does the checks internally.