CMakeLists.txt: Fix compilation problems on debian buster
Since 60fcadac (CMakeLists.txt: Check for more modern -std= flags as
well, 2019-07-25) we are using the most modern -std flags avaiable.

And since 4e259462 (.travis: Run tests on debian 10 as well, 2019-06-05)
we use debian buster for testing as well.

The combination of the two features broke compilation on debian buster
as it's gcc supports C++17 but we actually don't have valid C++17 code
as omniorb uses throw specifications, which are not valid anymore.

The error looks like

/usr/include/omniORB4/internal/orbOptions.h:95:57: error: ISO C++17 does not allow dynamic exception specifications
     virtual void visit(const char* value,Source source) throw (BadParam) = 0;

...

So let's not use C++17 and C++20 for now.