.travis: Fix installation of non-stock cppzmq
In 86174eea (Travis Tests: Avoid code duplication, 2019-06-05) we added
support for using the environment variable NON_STOCK_CPPZMQ to say if we
use the stock version or the non-stock version.

That commit got the logic wrong. Flip the naming to have now one less
negation.

STOCK_CPPZMQ=ON means that the system cppzmq is used, =OFF that we use
the latest cppzmq from the repo.