.travis: Avoid splitting up commands which should not be run separately
Docker creates for each RUN command a separate layer and that means that
it can resume it's operation from each RUN command when we create a new
image.

For package installation this is not what we want as we want to have the
latest packages.

Fix that. In addition this also makes the images smaller.