Use feature tests for checking if we have zmq::socket::disconnect
Since 9c02de30 (- Commit after merge with NextRelease branch at release 22135, 2013-02-22)
we check the libzmq version to decide if cppzmq's socket class has a
disconnect method or not.

This currently fails on debian wheezy.

As that is also the wrong approach in general, we now use a compilation
test in the cmake setup phase to determine if we have the disconnect method
or not. This is only done on non-Windows platforms, on Windows we assume
that we have a disconnect method (we currently always compile against a
cppzmq version which has it).

In addition the define got also renamed from ZMQ_HAS_DISCONNECT to
TANGO_ZMQ_HAS_DISCONNECT to make it clear that this is our define.