Avoid warning from double loop indexing
In the returned string vectors of get_polled_cmd/get_polled_attr we have
always two consecutive entries in the poll list, namely the
command/attribute name and its polling period.

So when we iterate over them we always increment the loop
counter/iterator by two. This is currently done with one inside the foor
statement and one at the end of the loop body.

This triggers with clang 9.2.0 the following warning

/home/firma/devel/cppTango/cppapi/server/dserverpoll.cpp:1591:7: warning: variable 's_ite' is incremented both in the loop header and in the loop body [-Wfor-loop-analysis]
                                ++s_ite;
                                  ^
/home/firma/devel/cppTango/cppapi/server/dserverpoll.cpp:1581:59: note: incremented here
                        for (s_ite = cmd_list.begin();s_ite < cmd_list.end();++s_ite)

and as that is also heavily confusing we now increment in the for
statement only.