.travis.yml: Fix exit code ignoring for failed tests
In 753c3f1a (.travis.yml: Fail travis tests fast, 2019-10-23) we added
support for failing the build immediately on error.

This worked as it should.

But in 75dc8a2e (.travis.yml: Add environment variable which allows to
skip running the tests, 2019-10-23) support for not executing the tests
was added. And that failed to return the exit code from the test, as it
returned on error true from the outer ||.

Let's rewrite it so that it fails early and allows to skip the tests.