appveyor.yml: Add inline deployment on tag push
Automate the deployment from appveyor.

We want to publish all artifacts as a new draft prelease on tag push
only.

For reference see https://www.appveyor.com/docs/deployment/github/ and
https://www.appveyor.com/docs/appveyor-yml/.

I have created a new github machine user named cppTango-bot for that.
And also a personal access token which was encrypted on the appveyor
site.