Add ABI/API compliance checking using github actions
Maintaining ABI/API compatibility in C/C++ is a difficult task.
Fortunately the abi-compliance-checker tool [1] can help examining the
current state and compare different versions.

The script check-ABI-API-compliance.sh allows to compare the current
revision with the revision from the reference branch (currently
tango-9-lts).

This script is now executed as github actions job, but can also be executed
locally. We choose github actions because that allows to store artifacts
for 90 days for each PR.

[1]: https://lvc.github.io/abi-compliance-checker/