configure/CMakeLists.txt: Enhance flags on *nix
Unfortunately we can not just use '-Og' as in d46dd5a8
(configure/CMakeLists.txt: Enhance and simplify compile flags on *nix,
2020-01-23) as our oldest GCC is 4.7 from debian wheezy.

So we still need to check if '-Og' is available.

We want to compile with '-Og' for abi compliance-checks [1] but we do
that with more recent GCC versions.

And we also add the relevant security compile options from
dpkg-buildflags. -fstack-protector strong is only added for the release
builds as that might interfere for debug builds.

[1]: https://lvc.github.io/abi-compliance-checker/