.travis.yml: Fix SONAR_SCANNER=ON failure
Currently the sonar scanner is broken since october
2019 according to [1].

It is not clear when and how it got broken.

But from the CI error message

[0K$ test ${SONAR_SCANNER} = "ON" &&
.travis/sonar.sh
Using SONAR_PROPERTIES=-Dsonar.pullrequest.base=tango-9-lts
-Dsonar.pullrequest.branch=reset-database-and-ds-for-each-test
-Dsonar.pullrequest.key=640
.travis/sonar.sh: line 11: sonar-scanner: command not found

it is at least clear that sonar-scanner is missing and not in the PATH.
So let's do that.

[1]: https://sonarcloud.io/dashboard?id=org.tango-controls%3Acpp-tango%3Atango-9-lts