.travis.yml: Fix SONAR_SCANNER=ON failure
Currently the sonar scanner is broken since october
2019 according to [1].

It is not clear when and how it got broken.

But from the CI error message

[0K$ test ${SONAR_SCANNER} = "ON" &&
.travis/sonar.sh
Using SONAR_PROPERTIES=-Dsonar.pullrequest.base=tango-9-lts
-Dsonar.pullrequest.branch=reset-database-and-ds-for-each-test
-Dsonar.pullrequest.key=640
.travis/sonar.sh: line 11: sonar-scanner: command not found

it is at least clear that sonar-scanner is missing and not in the PATH.

The following changes have been done in order to get it working:
- Move execution into debian 8 container
- Define SONAR_TOKEN in repository settings and pass it to the docker container
- Always exectute the build-wrapper if either COVERALLS or SONAR_SCANNER
is enabled. The reason is that sonar itself does not generate coverage
information, but we rely on the cmake coveralls support to generate that
for us
- Pass the SONAR_TOKEN to sonar-scanner

Due to how travis handles secrets the SONAR_TOKEN is not available in
external pull requests. Even encrypted environment variables are not
available on forks [2].

[1]: https://sonarcloud.io/dashboard?id=org.tango-controls%3Acpp-tango%3Atango-9-lts
[2]: https://docs.travis-ci.com/user/environment-variables/#defining-encrypted-variables-in-travisyml