Track event subscriptions by device proxy pointer
The pointer to device proxy used for event subscription is now stored in
EventSubscribeStruct (which is unique entry per subscription) instead of
EventCallBackStruct (which is shared between all subscriptions to the
same event).

Ability to associate subscriptions with device proxies is needed to
correctly unsubscribe from events inside device proxy destructor.