.travis/check-ABI-API-compliance.sh: Make it work
Turns out that the compliance check tool does not work in the current
setup.  The issue tracker [1] also suggests that we are not the only
one.

After fiddling with a lot of different flags and options of the
abi-dumper and the abi-compliance-checker tool, it turns out that diff
can help us here. We just diff the textual representation of the dump,
and if we have differences, it most likely broke.

The sed invocation removes the volatile parts of the dump with a fixed
string.

This also makes getting the html report as artifact obsolete.

[1]: https://github.com/lvc/abi-compliance-checker/issues