Adapt tests to single-core machines (Gitlab CI/CD)
Some tests exhibit different behavior on a single-core machines in
scenario where a call to poll_attribute is immediately followed by
a call to subscribe_event (for e.g. periodic events).

Most likely the issue is related to the polling algorithm and to the
detection of late events, but further analysis is required to confirm
that.